rest/media: Don't log `Requested file is too large`
When the remote resource is too large to be fetched, it's not an error we
can do anything about. Hence, this should go to the log as error or
warning, it's simply a common case of operation.

This patch removes these log messages:

```
2019-07-14 14:41:52,063 - synapse.http.matrixfederationclient - 874 - WARNING - GET-212- {GET-O-24} [matrix.org] Error reading response: 502: Requested file is too large > 10485760 bytes
2019-07-14 14:41:52,064 - synapse.rest.media.v1.media_repository - 390 - ERROR - GET-212- Failed to fetch remote media matrix.org/GPMcaQOeYjjWNTMFarGJfQBJ
Traceback (most recent call last):
  File "/usr/lib/python3/dist-packages/synapse/rest/media/v1/media_repository.py", line 373, in _download_remote_file
    "allow_remote": "false",
synapse.api.errors.SynapseError: 502: Requested file is too large > 10485760 bytes
```